third_party/blink/web_tests/fast/dom/Document/document-charset-expected.txt

Initial

 document.charset: KOI8-R

 document.characterSet: KOI8-R

 document.inputEncoding: KOI8-R

Setting charset to UTF-8... (expected to have no effect)

 document.charset: KOI8-R

 document.characterSet: KOI8-R

 document.inputEncoding: KOI8-R
